rename server to fullnode
