+++ src/lib.rs
@@ -38,7 +38,6 @@ pub mod request_processor;
 pub mod request_stage;
 pub mod result;
 pub mod rpu;
-pub mod server;
 pub mod signature;
 pub mod sigverify;
 pub mod sigverify_stage;
